+++ src/pas/plugins/headers/plugins.py
@@ -130,12 +130,19 @@ class HeaderPlugin(BasePlugin):
             return True
         if self.redirect_url:
             url = self.redirect_url
+            # url is expected to be a native string both on Py 2 and 3,
+            # but I have seen it as bytes on Py 3 in a traceback.
+            if six.PY3 and isinstance(url, bytes):
+                url = url.decode('utf-8')
             # If url is headerlogin, we want localhost:8080/Plone/headerlogin
             # and not localhost:8080/Plone/current-folder/headerlogin.
             # So relative from the site root.
             # Or from the navigation root, but that needs a context, which we don't have here.
             # Watch out for '//some.domain' as external redirect url.
             if '//' not  in url:
+                if not url.startswith('/'):
+                    # Avoid getting .../Ploneheaderlogin as url.
+                    url = '/' + url
                 url = api.portal.get().absolute_url() + url
             url = '{}?came_from={}'.format(url, request.URL)
             logger.warning('Redirecting to %s', url)